GameController functionality to handle game added to GameService, gameState had previous turn int added for easy handling of round. GameStateTransport was modified for creation sake.

+++ api/src/main/java/com/pokerface/pokerapi/game/GameController.java
@@ -60,7 +60,7 @@ public class GameController {
         }
     }
 
-    private GameStateTransport handleAction(long gameId, GameAction action, long playerId) {
+    private GameStateTransport handleAction(long gameId, GameAction action, int playerId) {
         GameStateTransport nextGameState = gameService.handleAction(gameId, action, playerId);
         messenger.convertAndSend("/messages/game/" + gameId, nextGameState);
